requirements.txt
llama-index-core
llama-index-readers-file
llama-index-llms-openai-like
llama-index-embeddings-huggingface
llama-index-vector-stores-faiss
faiss-cpu
rank_bm25
sentence-transformers
